final_tool/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# final_tool/deeplab/deeplab_tools.py: 7
# final_tool/facepic_generator.py: 7,12
Pillow == 5.4.1

# final_tool/deeplab/deeplab_tools.py: 5
# final_tool/dnnlib/tflib/autosummary.py: 26
# final_tool/dnnlib/tflib/network.py: 15
# final_tool/dnnlib/tflib/optimizer.py: 10
# final_tool/dnnlib/tflib/tfutil.py: 11
# final_tool/dnnlib/util.py: 14
# final_tool/facepic_generator.py: 9
numpy == 1.16.2

# final_tool/deeplab/deeplab_tools.py: 4
# final_tool/facepic_generator.py: 6
opencv_python == 4.2.0.32

# final_tool/dnnlib/util.py: 22
requests == 2.21.0

# final_tool/dnnlib/tflib/autosummary.py: 28,29
tensorboard == 1.15

# final_tool/deeplab/deeplab_tools.py: 6
# final_tool/dnnlib/tflib/autosummary.py: 27
# final_tool/dnnlib/tflib/network.py: 16
# final_tool/dnnlib/tflib/optimizer.py: 24,27
# final_tool/dnnlib/tflib/tfutil.py: 12
tensorflow == 1.15

# final_tool/facepic_generator.py: 11
tqdm == 4.31.1
